+++ packages/flutter_query/lib/flutter_query.dart
@@ -55,6 +55,7 @@ export 'src/hooks/hooks.dart'
         useQueryClient,
         useQuery,
         useMutation,
+        useMutationState,
         useInfiniteQuery,
         useIsFetching,
         useIsMutating;