+++ packages/flutter_query/lib/src/hooks/hooks.dart
@@ -2,5 +2,6 @@ export 'use_infinite_query.dart';
 export 'use_is_fetching.dart';
 export 'use_is_mutating.dart';
 export 'use_mutation.dart';
+export 'use_mutation_state.dart';
 export 'use_query.dart';
 export 'use_query_client.dart';